+++ website/addons/s3/templates/s3_user_settings.mako
@@ -42,18 +42,18 @@
     <script type="text/javascript">
 
         $('#s3RemoveAccess').on('click', function() {
-            bootbox.confirm(
-                'Are you sure you want to delete your Amazon Simple Storage Service access key? This will ' +
-                    'revoke access to Amazon S3 for all projects you have authorized ' +
-                    'and delete your access token from Amazon S3. Your OSF collaborators ' +
-                    'will not be able to write to Amazon S3 buckets or view private buckets ' +
-                    'that you have authorized.',
-                function(result) {
-                    if (result) {
+            bootbox.confirm({
+                title: 'Remove access key?',
+                message: 'Are you sure you want to delete your Amazon Simple Storage Service access key? ' +
+                        'This will revoke access to Amazon S3 for all projects you have authorized and ' +
+                        'delete your access token from Amazon S3. Your OSF collaborators will not be able ' +
+                        'to write to Amazon S3 buckets or view private buckets that you have authorized.',
+                callback: function(result) {
+                    if(result) {
                         deleteToken();
                     }
                 }
-            )
+            });
         });
 
         function deleteToken() {
